+++ chronobio/game/loan.py
@@ -6,13 +6,13 @@ class Loan:
         self.amount = amount
         self.start_day = start_day
 
-    def month_cost(self: "Loan", day: int) -> int:
+    def month_cost(self: "Loan", day: int) -> float:
         if not (self.start_day <= day < self.start_day + LOAN_DURATION_IN_MONTHS * 30):
             return 0
         return self.amount * LOAN_INTEREST / LOAN_DURATION_IN_MONTHS
 
-    def remaining_cost(self: "Loan", day: int) -> int:
-        total = 0
+    def remaining_cost(self: "Loan", day: int) -> float:
+        total = 0.0
         for month in range(LOAN_DURATION_IN_MONTHS):
             cost = self.month_cost(day=day + month * 30)
             if not cost:
